⚠ MAJOR COMMIT: Now shifted to dist styled package
Udemyscraper should be installed as package. (pip install -e .)
Refractured the main file into 5 files under udemyscraper dir
A lot of bug fixes
