--- setup.cfg
@@ -1,24 +1,0 @@
-[metadata]
-name = udemyscraper-sortedcord
-version = 0.7.0
-author = Aditya Gupta
-author_email = divyashgupta2005@gmail.com
-description = A Udemy Course Scraper built with beautiful soup, that fetches udemy course information. Get udemy course information and convert it to json, csv or xml file, without authentication.
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/sortedcord/udemy-web-scraper
-project_urls =
-    Bug Tracker = https://github.com/sortedcord/udemy-web-scraper/issues
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: GPL-3.0 License
-    Operating System :: OS Independent
-
-[options]
-package_dir =
-    = src
-packages = find:
-python_requires = >=3.8
-
-[options.packages.find]
-where = src
