+++ udemyscraper/metadata.py
@@ -1,0 +1,7 @@
+__package_name__ = 'udemyscraper'
+__version__ = "0.7.4"
+__author__ = 'Aditya Gupta'
+__author_email__ = 'divyashgupta2005@gmail.com'
+__description__ = """A Udemy Course Scraper built with beautiful soup, \
+that fetches udemy course information. Get udemy course information \
+and convert it to json, csv or xml file, without authentication."""
